+++ empire/Godeps/_workspace/src/github.com/remind101/pkg/httpx/middleware/logger.go
@@ -6,8 +6,8 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/remind101/empire/empire/pkg/httpx"
-	"github.com/remind101/empire/empire/pkg/logger"
+	"github.com/remind101/pkg/httpx"
+	"github.com/remind101/pkg/logger"
 	"golang.org/x/net/context"
 )
