+++ empire/Godeps/_workspace/src/github.com/remind101/pkg/httpx/middleware/recovery.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/remind101/empire/empire/pkg/httpx"
-	"github.com/remind101/empire/empire/pkg/reporter"
+	"github.com/remind101/pkg/httpx"
+	"github.com/remind101/pkg/reporter"
 	"golang.org/x/net/context"
 )
