+++ empire/Godeps/_workspace/src/github.com/remind101/pkg/logger/example_test.go
