+++ empire/Godeps/_workspace/src/github.com/remind101/pkg/logger/logger_test.go
