+++ empire/Godeps/_workspace/src/github.com/remind101/pkg/reporter/hb/generator.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"runtime"
 
-	"github.com/remind101/empire/empire/pkg/httpx"
+	"github.com/remind101/pkg/httpx"
 	"golang.org/x/net/context"
 )
