+++ empire/Godeps/_workspace/src/github.com/remind101/pkg/reporter/logger_test.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"testing"
 
-	"github.com/remind101/empire/empire/pkg/logger"
+	"github.com/remind101/pkg/logger"
 	"golang.org/x/net/context"
 )
