--- empire/jobs.go
@@ -1,1 +1,0 @@
-package empire
